Не рабочая версия - проба2

+++ eSh/eScheme/eScheme/ISetValue.vb
@@ -1,3 +1,3 @@
 ﻿Public Interface ISetValue
-	Sub SetValue(value As Integer)
+    Sub SetValue(value As Single)
 End Interface
